MB-41165: Modify StatChecker to be able to check post-warmup
Currently attempting to use the StatChecker and it's subclasses across
a warmup will result in a segfault as it holds a reference to the
VBucket object. I'd like to use it in this way in a future patch to
test that we have persisted new counts during compaction correctly.
Modify the StatChecker to instead take a reference to the VBucketPtr
(which the caller must reset after the warmup) to facilitate this test.

Change-Id: Ib44883a81d6402f6f25d282660bdf4301bed2591
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141668
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>