MB-43156: Update waketime in FollyExecutorPool::wake()
If a task managed by an ExecutorPool is scheduled to run immediately
via wake(), it's wakeTime - the time it was desired to run - should be
reset to "now" - given that's the new time for it to run. waketime is
then used to calculate the queue time of the task - how long the task
has been waiting to run, which is exposed via cbstats scheduler.

However the FollyExecutorPool implementation fails to update the
waketime when wake() is called; leaving at its previous value. This
can give inaccurate values for queue time, as the the queue time is
reported as zero (capped from a large negative value:

    now() - waketime.

Change-Id: Ic13a53065890f0bdb1bfa7daa914ba1cca11af11
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141701
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>