MB-41321: Clean-up container for multiple 'drop collections'
The DroppedCollections container can have multiple entries
for a collection (if a resurrection of a collection occurs).

The clean-up which is driven by the flusher after a successful
commit only calls back into the Manifest once with the highest
drop event which was flushed.

The remove function is now updated so that it erases all entries
below or equal to the given seqno, ensuring full clean-up.

Change-Id: I1c7724c5192717e9c93cecf951046f668375764b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141794
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>