MB-43015: Don't acquire Configuration mutex in decodeWithMetaOptions()
EventuallyPersistentEngine::decodeWithMetaOptions() accesses the
conflict resolutoin mode via the Configuration, which is slow as
Configuration is actually implemented as a map requiring a lock to
access.

When measuring XDCR initial replication throughput at the destination
(i.e. SET_WITH_META throughput), approx 5% of worker thread runtime is
in configuration.getConflictResolutionType().

Avoid this by adding a conflictResolutionMode member variable to
EPEngine and using that instead.

Change-Id: I769ff6e364fe397109dd54e0cdbca6d2924f6330
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141800
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>