[Refactor] Move creation of openssl SSL structures
Previously we created the SSL structure as part of the Connection
structure, but this patch moves the creation of the structure
before we try to allocate a Connection struct.

This is a preparation for the refactor to move off the interface
descriptors inside memcached.json and put all connection properties
inside the network configuration manager. Given that we accept
all connections in a single thread (the dispatch thread) we could
move creation of the SSL structure to this thread and get rid of
the extra complexity related to locking and reconfigure of the
TLS properties.

Change-Id: Ie643edca3ec68f931f131c964a6e26712e7a916b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141942
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>