MB-42969: engine_testapp: Run cleanup functions if test case throws
If the test throws an exception during exeuction (including a
TestExpectationFailed exception which is used to explicilty flag
expectation failures) then catch that exception earlier, and ensure
all test / bucket cleanup functioins are executed.

This should ensure that all objects are correctly cleaned up /
destructed and subsequent tests have a clean environment to run in,
after a test failure.

Change-Id: I925038d73cde3872a30a8b55cd815938495b18de
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142114
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>