MB-43238: Take a FindUpdateResult in VB::processExpiredItem
In VBucket::processExpiredItem we must also consider if there is an
incomplete prepare before attempting to expire the committed item
(we will never expire a pending one). To do this we should pass the
FindUpdateResult to processExpiredItem so that we can make the change
in a future patch.

This also requires a change to FindForUpdate to take a reference to
the HashTable so that we can reuse the StoredValue selection logic in
findForRead in FindUpdateResult.

Change-Id: Iaf5170cf94c4111b0594af00e8fa32e5c5739674
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142165
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>