[Refactor] Modify client cert auth logic
Modify the way we do client certificate order:

1. Pick out the setting from the SSL object

The SSL object contains verification flags that OpenSSL will use as
part of connection establishment, and we may continue to use these
flags to determine how we shall continue to perform the authentication.
By using these flags (which is stored inside the SSL object) we don't
need to acquire any locks in order to figure out the configured mode.

2. Only look at the peer certificate if we have to

If the configured mode is disabled we don't need to pick out the
peer certificate (and perform memory allocation).

2.1 Only try to look up the user if the peer provided a certificate

If the configuration is Enabled or Mandatory _and_ the client
provided a certificate then try to look up a user from the provided
certificate and report the appropriate error message.

Also fix the error reported logged from bufferevent where we shut
down the connection when we see unrecoverable erros. Previously we
printed out the socket error, but this may be from when OpenSSL
refuse to open the connection (for instance missing peer certificatei
if the mode is set to mandatory) causing the log to look like:

INFO 83: Connection::on_event: Socket error: Success

it looks better with:

INFO 83: Connection::on_event: unrecoverable error encountered

Change-Id: I67d3ddb4f9c9274509f124badcab3735b00ea278
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142167
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>