MB-43331: Inc expiry stat on deletion of expired item
We should increment the expiry stat if we expire an item when we
attempt to delete it.

Additionally in this patch it was noticed that we increment the stat if
we attempt to expire an item that does not exist in either the HashTable
or the BloomFilter for FullEviction buckets. This is an odd case as we
should always find the key in the BloomFilter unless compaction has
recently run and created a new BloomFilter without it. If compaction
has done so then it means either the item has been deleted for the
metadata purge interval or was dropped as part of a collection drop.
The latter case is the more interesting one here. We could find that
an item needs expiring in the PagingVisitor and add it to the list to
be expired on completion of the iteration. The collection could then
run and compaction could drop it before attempting to expire the item.
In this case the code would increment the expiry stat. Doing this is
of questionable value so this patch has tweaked that behaviour to not
increment the stat.

Change-Id: I741e3962033d991c271b50fd64cb13c07e144ebb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142263
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>