MB-43300: Don't iterate all collections when getting the stats of one
cmd_stat which requests the 'summary' stats of one collection visits
all active vbuckets and maintains a temporary data structure with
stats for every collection. If the system is configured with many
collections (e.g. the max of 1k) the cmd_stat request becomes measurably
slow (see table at end of commit message).

Similarly the stat request for a scope can be affected in a similar way
when the scope is only a small subset of the total collection set.

This patch re-factors the one-collection and one-scope stats "summary"
code, i.e. the handling for cmd_stat with key:

* collections name
* collections-byid id
* scopes name
* scopes-byid id

The re-factoring adds a new vbucket visitor and supporting code so
that when stats are requested only the collections of interest have
stats gathered. The visitor can be configured with 1 or many
collections.

In the one-collection case we have a vector with one entry, and in the
one-scope case we have the collections that belong to the scope, in both
cases the visitor never has to operate on the entire collection set,
just the collections in the input vector.

The re-factoring shows a measurable improvement as follows.

A simple timing exercise was done using mc_bin_client.py to invoke
cmd_stat collections-byid. A 2 node cluster was used with 1,000
collections configured. A single connection then looped n times
invoking cmd_stat against a different collection, the timing of
the loop is as follows for before this patch and with.

n    | time before (seconds) | time with (seconds)
1    | 0.0077                | 0.0008
10   | 0.0765                | 0.0056
100  | 0.7713                | 0.0636
1000 | 7.7914                | 0.6801

Change-Id: I56d1d92d2862b2b4c0601de7249abde83ffee53c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142826
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>