MB-43345: Fix DCP_CONTROL msg v7_dcp_status_codes
This patch fixes v7 dcp status codes enablement on the DcpConsumers
end. Before this patch we incorrectly enabled the use of v7 dcp status
codes on the DcpConsumer end if the consumer had successful send the
DCP_CONTROL message with key="v7_dcp_status_codes" and value="true".
This would main that when a pre 7.0.0 returned ENGINE_EINVAL we would
ignore it and keep using v7 dcp status codes.

To fix this we now correctly negotiate if the consumer should send v7
DCP status codes by waiting for the DCP producer to reply with
ENGINE_SUCCESS.

Change-Id: I0baa08306964e82d2338a113d89294232f983eec
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142967
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>