MB-42737: Don't terminate memcached when reconfig of prometheus fails
Terminate memcached if we fail to open the prometheus port during
startup, but if some error happens during reconfigure we should
just log it instead of terminating the process.

The current reconfigure scheme of interfaces is risky as we may
not be able to open up ports. This is all addressed in MB-39620
where we provide commands to define/delete interfaces so that
the calling process knows which ports we listen on.

Killing memcached for a reconfigure failure will cause data loss
of all items not persisted to disk/replicated and make memcached
enter a warmup phase. ns_server may work around the problem today
by trying to reload the configuration and we'll retry binding to
the port. If that continues to fail it should move to a different
port (or even better: use an ephemeral port)

Change-Id: If29159d94f06f96d132e08de16eee6cfb73d4b00
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142971
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>