MB-43480: Disable FollyExecutorPool dynamic thread creation
When dynamic thread creation is enabled, folly will join threads
that have been idle for a (by default) 1 minute. Folly will
re-create threads as needed up to the maximum we define. This
recycling of threads is problematic for at least AtomicQueue
in that it assumes a fixed number of threads (up to a maximum of
500). With dynamic threads enabled AtomicQueue no longer
functions correctly and eventually throws exceptions once its
internal thread counter exceed the maximum.

Change-Id: I6c70abcbf8b21096def08dec331eecd837df1c26
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142985
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>