MB-43373: Abort memcached if exception thrown from GlobalTask::run
As seen in MB-43348, if an exception is thrown (and uncaught) inside a
GlobalTask::run() method when run on a background thread via the
FollyExecutorPool, Folly catches the exception (and logs an error
message). This is in contrast to CB3ExecutorPool, which doesn't catch
exceptions in it's background threads and hence the program will be
terminated.

While such uncaught exceptions are "incorrect", In the case of
FollyExecutorPool the consequence is that the task can become "stuck"
and no longer getting scheduled. In the case of MB-43348, an exception
was thrown during Flusher::run(), which meant the various state
updates to when next run the Flusher were not updated, and Flusher
task was hung.

To ensure consistent behavour between CB3 and Folly ExecutorPools,
mark GlobalTask::run() as 'noexcept', so uncaught exceptions are
guaranteed to terminate the program.

Change-Id: I5da2f2ba7acf7248d6f8132cd4fecf5712540d2c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143046
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>