MB-37920: Don't update any cached vbstate member if persist not needed
Currently in KVStore::updateCachedVBState() we determine if persisting a
new vbstate is necessary by checking if any value has changed for a subset
of members of vbucket_state.

The interesting part is that, if persistence is NOT required, then we
proceed to updating part of the cached vbstate anyway. That introduces a
potential logic error where the cached vbstate is not completely aligned
with the actual vbstate on disk.

I name it a "potential" logic error because that appears to never happen
actually. That because:
 1) updateCachedVBState() is executed only in the code-path for when we
    have only a new vbstate to persist (ie, no mutations)
 2) the subset of members of vbucket_state that are updated anyway are:
     - highSeqno
     - maxDeletedSeqno
     - maxCas
     - hlcCasEpochSeqno
     - mightContainXattrs
    , which gets all updated only when persisting mutations. Plus:
     - lastSnapStart
     - lastSnapEnd
     - checkpointType
    , which define a snapshot. Given that our logic at flushVBucket() is
    "never appear to be in a snapshot on disk if there is no muation for
    it", then in this case any wrong update to the cached vbstate is
    prevented by our logic in EPBucket::flushVBucket().

With this change updateCachedVBState() doesn't touch any value in cached
vbstate is the new vbstate is not to be persisted. That removes any
chance of inconsistency between cached/on-disk vbstate previously
introduced by the function.

Change-Id: Ieb5ab279322cf29d5e02ff192fbd7c994094e866
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143027
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>