MB-37920: Update cached vbstate only if flush succeeds
Core patch for this MB. Currently we may update the cached vbstate if
flush fails, which is fixed here.

Note that this MB covers only the case of empty flush batches (ie, no
mutations, only a new vbstate persisted). Similar issues for non-empty
batches have been already spotted and fixed, so we appear to be fine in
that area.

Change-Id: I76bbb060fa6129574ae0501423cfc52c56c1a3be
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142987
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>