MB-42979: Fix intermittent failure of cursor dropping test
test_dcp_cursor_dropping_backfill makes assertions about the snapshot
start seqnos, which depend on the _end_ of the preceding backfill.

Waiting for items to be persisted ensures the backfill ends at a known
seqno for persistent buckets, but ephemeral backfill may end at a seqno
lower than the high seqno.

The test failed when the loaded items happened to be in multiple
checkpoints, and were not all removed from memory before the backfill
began. The ephemeral backfill would end at the last item not in memory.

Fix by having the test wait for all stored items to be removed from the
checkpoint manager.

Change-Id: Ia7ef20b10a8c2dfd9970263b98799c5d3b5e8a2e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143047
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>