MB-43526: Reset dirtyQueueAge at rollback
Currently at rollback we reset dirtyQueueSize but we do not reset
dirtyQueueAge. These two stats are typically linked (i.e. we set on
and the other at the same time). Not resetting dirtyQueueAge here
means that a queue_age calculation will subsequently be wrong as it
will include the cumulated dirtyQueueAge of any items that were still
unpersisted at rollback.

Change-Id: Ieea76592548bf6b2466854ca0309899a4e03e8fb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143128
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>