Rename KVStore::getVBucketState to getCachedVBucketState
The function returns the vBucket state from the cache but that's not
particularly clear from the name. In a future patch I'd also like to add
a KVStore::getPersistedVBucketState function to return the on disk
state. This should make the KVStore API more intuitive and will allow us
to read the persisted vbucket_state of vBuckets without casting the
KVStore to the child type.

Change-Id: I84a258b49040d644bec93dbd566eee3603b1b49c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143189
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>