MB-39292: 1/n VB::Manifest holds shared_ptr to the manager
This is 1 of a series of patches that improve collection metadata
handling with the final goal of detecting when scopes or collections
have a legal or illegal change of name. Prior to this series of patches
there is no easy way for the VB collection update logic to access the
name of a collection.

Some of the aims of the patch series:

* To keep the VB collection update logic reasonably similar to how
  it worked before this series. Specifically avoid any need to to
  'ewouldblock' patterns during DCP induced takeover or direct
  set-vbstate.

* To avoid further memory consumption (certainly scaling with
  collections * vbuckets). Collections and scopes *generally* have
  the same name, so part of this series is to keep 1 bucket copy rather
  than 1 vbucket copy (and a name could be 251 bytes).

This series begin here by reorganising the Collections::Manager
ownership. The Manager will be the object that stores the name (and
even more), and as such the KVBucket owned Manager needs to be
accessible from a VB::Manifest.

This isn't as simple as giving a reference to the 'unique' Manager
because of the way that VBuckets can outlive the KVBucket. This
certainly happens in VBucket deferred deletion where a deletion task
can run after KVBucket destruction.

This is why we switch from unique to shared ownership, allowing the
VB::Manifest to access the Manager, even after KVBucket destruction.

Change-Id: Ib037cb178848fe9cdcb8b3dbe7ae7267738f0f3d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144491
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>