MB-43529: Correct dirtyQueueAge underflow when we de-dupe
To try to ensure that we can't underflow the dirtyQueueAge value we
should make it a NonNegativeCounter. This highlighted an interesting
case where we could in fact make it underflow if we de-duped an Item
in the CheckpointManager as flushing would then attempt to decrement
the queueTime of the newer version of the item (rather than the
original queueTime). Correct this by having the de-dupe code overwrite
the queueTime if the Item has not yet been persisted.

Change-Id: I372bc6f85802e8e826c4106c2cb821ba9a11416c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143278
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>