MB-43595: Prefix the engine name into the magma logger name
We should add the engine name to all magma log messages to work out
which engine log messages belong to. In the Bucketlogger this is done
on the fly by the logging function using the engine thread local. It's
not possible to do this in the same way for magma as the spdlog log
functions that we override in the BucketLogger are non-virtual and
magma can't add a dependency to ep engine to use the BucketLogger.

Prefix the engine name into the logger name for magma.

Change-Id: I14c512bafcbf5b2880315a567fe6c4eaa6feebba
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143334
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>