MB-43199: Capture and notify cookie(s) for all compaction requests
MB-41782 updated compaction task management so there should be just
one task in existence per vbucket, be it scheduled from a memcached
command or a collection drop. In the case where a command already
exists, the scheduling code 'pokes' the existing task with
information for the new schedule request, the task itself will
either run with the new request or run again for the new request.
In this case any cookie associated with the memcached command is
not passed into the existing task meaning depending on schedule
request ordering, there may never be a notifyIOComplete callback.

This commit updates the code which updates an existing task to
ensure the task is also given the cookie, so it can notify when
done. As there could be multiple requests from multiple connections,
the CompactTask can store multiple cookies to be notified.

There is though one issue to be addressed, and that is when multiple
requests occur with differing configs, compaction aims to run
the minimum number of times (as it can be expensive) so there is
currently no guarantee that all of the requested options will be
respected. E.g. ns_server may request compaction with a 'normal'
configuration and then a command with ""

For example a request for compaction with "drop_deletes=true" which
is followed by a request for compaction with "drop_deletes=false" may
mean "drop_deletes=true" is never applied to the vbucket.

Change-Id: I6affcdccb1352cfc261303e3a6c87f3c684afec2
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143380
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>