MB-43599: Pass correct StoredValue to processExpiredItem
When we call processExpiredItem on an item that does not exist in the
HashTable for a full eviction bucket we currently crash in
VBucket::processExpiredItem as the StoredValue passed in as part of
the HashTable result is not present. This is the case because we create
a new StoredValue to expire the item but do not update the HashTable
result.

Update the HashTable result to point to the newly created StoredValue.

Change-Id: If2ddeff8c7e551766b415cb9378c6b7cbf7c6b79
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143357
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>