MB-43199: Run compaction with a config representing multiple requests
If multiple compaction requests occurred before the task runs, the
config of the last schedule request is what will be used in the run.

For example a user may issue a manual compaction (via cbepctl) which
requests all deletes are dropped, whilst ns_server requests a normal
compaction. The user's command may return success, yet a compaction
with drop_deletes=true may never of executed. Prior to MB-41782
the user's request would of been honoured because separate compactions
would of been scheduled.

This commit uses a merge function on the CompactionConfig so that
all schedules 'merge' the requested config into the scheduled task's
config. When the CompactionTask runs it now has a config which
respects all requests.

Change-Id: I6e2d373e0c6f36dd946db06c3d6953a7481c1eae
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143381
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>