MB-43485 magma-kvstore: Prevent kvstore not-exists error from getting ignored
Magma Get() APIs return status=ok when the API call is successful. The found
flag specifically indicates whether the key is found or not. An error
code check in readLocalDoc() was ignoring not-exists error code and
treating it as successful read. This results in not able to distinguish
between a key not found vs the kvstore does not exist.

Change-Id: I3b23f2eba5766abb63902c2475dde4bdc29949b0
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143409
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>