MB-42568: [Ephe] Correctly update num-deleted-item in updateStoredValue
In EphemeralVBucket::updateStoredValue we try to update an existing OSV
and move it to the end of the SeqList if possible.
That may be prevented by range-reads that cover that OSV. In that case
we add a new stored value at the end of the SeqList and leave the old
OSV at its place.

Before this patch we used to always update the SeqList::numDeletedItems
counter based on the semantic of an "update SV", also for when the "add
SV" logic is executed.

That led to over-decrementing the num of deletions, which underflows the
counter in some scenarios.

Change-Id: Ib5c8910d68a4edcc0da1bf16e62e0acb51c70894
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143512
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>