MB-43205: Introduce the allow_sanitize_value_in_deletion config param
This is a renaming of the previous allow_del_with_meta_prune_user_data
config param.

As per its old name, that was a switch for enabling/disabling
del_with_meta operating in "sanitizer" mode rather than enforcing a
strict validation and failing the operation in the case of invalid
payload.
Now we want to extend the same behaviour to DCP_DELETE/DCP_PREPARE.

That is because in 6.6.0 we have introduced some stricter validation
on deletion payloads that may fail the operation. That may happen
mainly in the case of pre-6.6 to 6.6 offline upgrade.

Under this MB we want to introduce the possibility to set a replica
in "sanitizer" mode, the same as we already do at del_with_meta. That
way, DCP_DELETE/DCP_PREPARE will just remove any invalid body in the
payload rather that failing.

Change-Id: Ia9faff48de3a51a77d367961b45c41ed45c609d1
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143656
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>