Create DCP consumer / producer without holding the lock
Previously we grabbed a write lock to the underlying map and
checked if we already had the cookie registered (which would
_ALWAYS_ fail as the core won't allow DCP open on a connection
which is already marked as DCP). After that it would check
if we had a connection with the same logical name before we
would create the object and insert it into the map.

We only need the lock while we check for the name and insert
the new stream.

Remove the code to check for an existing consumer/producer
for the cookie and the unit tests for that (it is tested
via DcpTest::CantDcpOpenTwice that the second DCP open
won't reach the underlying engine and is refused from
the packet validators)

Change-Id: I70a1f4a3217baa07405b2156d2bb84d3931c8c11
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143688
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>