MB-43681 magma-kvstore: Fix incorrect kvstore revision update
When a new kvstore is created, the vbstate is written as the first
update to the kvstore. If the process crashes just before vbstate write
is complete, the vbstate document will be missing. When the kvstore
constructor tries to populate the vbstate during warmup, it uses
a default vbstate if the document does not exist. The default revision
used for the kvstore is 0. If a kvstore is already created with revision
1 and no vbstate document written yet, the vbstate generated for the
kvstore during warmup will be incorrect. This results in a stale
kvstore version to exist in the magma db.

We have to correctly populate the kvstore revision even if a vbstate
document does not exist by reading the revision from magma.

Change-Id: I3c91bd5c2372aef9a76f7afed02cbb5ab6d385cf
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143775
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>