MB-43622: Fix race between DCP Open and scheduleDcpStep
Previously there was a possibility of a race for DCP
connections by the fact that DCP Open would create the handler
but it wouldn't set the state for the connection to DCP until
the method returned. In the mean time the cookie was available
for DCP to try to schedule a step.

This change sets the connection up as a DCP connection as part
of creating the the ConnHandler object. Unfortunately we can't
reserve the cookie as part of the constructor without a massive
change working through all of the unit tests as they don't clean
up the reference by calling release.

To work around the race condition we'll add an extra ref count
to the cookie before calling DCP Open

Change-Id: Ifa9b87b984af6be53934c8100e1a8d584c423c13
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143492
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>