MB-43745: Check for collection existence before updating stats
Ephemeral pageOut uses a VBucket::doCollectionsStats that can
result in an exception if the collection of the paged item
no longer exists.

This is fixed by moving the 'ReadHandle' variant of
doCollectionStats to ephemeral (as that's the only user, in
the pageOut function). Now as Item's are checked for eligibility
for paging, their collection existence is also checked. No
collection means no page-out.

Change-Id: I91eb966b62311d1ff62a6d4886cef4e1fd8bfcf5
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143802
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>