MB-43373: Abort memcached if exception thrown from GlobalTask::run v2
As seen in MB-43348, if an exception is thrown (and uncaught) inside a
GlobalTask::run() method when run on a background thread via the
FollyExecutorPool, Folly catches the exception (and logs an error
message). This is in contrast to CB3ExecutorPool, which doesn't catch
exceptions in it's background threads and hence the program will be
terminated.

While such uncaught exceptions are "incorrect", In the case of
FollyExecutorPool the consequence is that the task can become "stuck"
and no longer getting scheduled. In the case of MB-43348, an exception
was thrown during Flusher::run(), which meant the various state
updates to when next run the Flusher were not updated, and Flusher
task was hung.

To ensure consistent behavour between CB3 and Folly ExecutorPools, add
add a catch block at GlobalTask, which prints the exception and then
calls std::terminate. This will be similar to how uncaught exceptions
on the front-end threads are handled - it_won't_ give a backtrace all
the way back to where the exception was first thrown.

(Note: Previous attempt to fix this MB (f5fc8e7bb) changed
GlobalTask::run virtual method to 'noexcept', which triggers
std::terminate if any uncaught exceptions occur. However, that had the
downside of leaving a potentially misleading backtrace in the
minidump).

Change-Id: I5dcacb67db61d7b9b5dc2defadac4e04f7d5ac49
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143807
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>