MB-43708: Add proper wouldblock support to setCollections
The memcached function, collections_set_manifest_executor wasn't
structured to handle 'wouldblock' properly, and as such failure
of the background task meant the task resumed the ep-engine
function which leads to access of deleted data.

A second issue was spotted in that a failed setCollections didn't
clean-up some of the manager state, so subsequent setCollections
calls would return ebusy (as if the previous call was still
in-progress).

Change-Id: I775a3662e2902d67a4195efc3a6d0043b3ed044d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/143822
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>