MB-39292: 2/n Add a 'SharedMetaDataTable'
To be able to correctly identify legal or illegal changes to a
collection's metadata we need a location to store and access
the metadata (primarily the name, but other metadata will be included).

To avoid unnecessary memory increases, we need a place in memory where
the VB::ManifestEntry objects will be able to 'reference' the and not
own it.

This patch introduces a custom container type for the storage and
retrieval of the metadata to be associated with collections (and
scopes).

The intention is that the name, scope-id and ttl of a collection will
not be stored in the VB::ManifestEntry, but referenced.

This container deliberately does not make use of std::shared_ptr as
memory profiling during the development already highlights that
collections/scopes can give a noticeable increase to mem_used. With
shared_ptr holding two pointers and a desire to decrease the
collections mem_used foot-print the container uses our own
SingleThreadedRCPtr. The use of SingleThreadedRCPtr allows
VB::ManifestEntry to reduce in-size by scope-id/ttl, benefit from
the later inclusion of the name and only then increase by 1
sizeof(pointer). This results in a net reduction memory required to
store each vbucket's collection data.

The container uses a multimap for storage because we need to support
the possibility that different id->meta mappings can co-exist.

E.g.
* vbucket 0 has collection:x with name:fruit
* vbucket 1 has collection:x with name:vegetable

This can happen in failure circumstances when the vbuckets see
different state time-lines (which is intended to be later force
corrected).

Note the SharedMetaDataTable will be used by VB::Manifest for scope
name storage too, hence the template parameters for key and value.

Change-Id: I7c7ded00827bf4fa1978668a1ffaebe908749dc6
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144492
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>